tests/wpt/tests/svg/painting/parsing/stroke-opacity-valid.svg

<?xml version="1.0" encoding="UTF-8"?>
<svg xmlns="http://www.w3.org/2000/svg"
     xmlns:h="http://www.w3.org/1999/xhtml"
     width="800px" height="800px">
  <title>SVG Painting: parsing stroke-opacity with valid values</title>
  <metadata>
    <h:link rel="help" href="https://svgwg.org/svg2-draft/painting.html#StrokeOpacityProperty"/>
    <h:meta name="assert" content="stroke-opacity supports the full grammar 'alpha-value'."/>
  </metadata>
  <g id="target"></g>
  <h:script src="/resources/testharness.js"/>
  <h:script src="/resources/testharnessreport.js"/>
  <h:script src="/css/support/parsing-testcommon.js"/>
  <script><![CDATA[

test_valid_value("stroke-opacity", "-1");
test_valid_value("stroke-opacity", "0.5");
test_valid_value("stroke-opacity", "3");
test_valid_value("stroke-opacity", "-100%", "-1");
test_valid_value("stroke-opacity", "50%", "0.5");
test_valid_value("stroke-opacity", "300%", "3");

  ]]></script>
</svg>
