LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/stroke-opacity-valid-expected.txt


PASS e.style['stroke-opacity'] = "-1" should set the property value 
PASS e.style['stroke-opacity'] = "0.5" should set the property value 
PASS e.style['stroke-opacity'] = "3" should set the property value 
FAIL e.style['stroke-opacity'] = "-100%" should set the property value assert_equals: serialization should be canonical expected "-1" but got "-100%"
FAIL e.style['stroke-opacity'] = "50%" should set the property value assert_equals: serialization should be canonical expected "0.5" but got "50%"
FAIL e.style['stroke-opacity'] = "300%" should set the property value assert_equals: serialization should be canonical expected "3" but got "300%"
